+++ src/Components/Display/Bulle/UpBulle.tsx
@@ -1,52 +1,53 @@
-import * as React from "react"
+import * as React from 'react'
 import * as cn from 'classnames';
-
-import SvgIcon from "../SvgIcon/index"
-import { IconName } from "../../../Common/theming/icons";
-import { BulleStyle, IconStyle, ValueStyle, MessageStyle, ChildrenStyle } from "./styles";
-import UpBox from "../../Containers/Box";
-import withTheme, { WithThemeProps } from "../../../Common/theming/withTheme";
-
+import SvgIcon from '../SvgIcon/index'
+import { BulleStyle, IconStyle, ValueStyle, MessageStyle, ChildrenStyle } from './styles';
+import UpBox from '../../Containers/Box';
+import withTheme, { WithThemeProps } from '../../../Common/theming/withTheme';
 import defaultTheme from '../../../Common/theming';
+import { UpBulleProps } from './types';
 
-export interface UpBulleProps {
-    backgroundImage: string;
-    message: string;
-    icon: IconName;
-    value: number;
-    className?: string;
-}
+const UpBulle: React.FunctionComponent<UpBulleProps & WithThemeProps> = props => {
 
-class UpBulle extends React.Component<UpBulleProps & WithThemeProps>{
-    
-    public static defaultProps: Partial<UpBulleProps> & WithThemeProps = {
-        theme: defaultTheme,
-    }
+    const {
+        className,
+        backgroundImage,
+        icon,
+        value,
+        message,
+        children
+    } = props;
 
-    constructor(p, c) {
-        super(p, c);
-    }
+    return (
+        <div
+            className={cn(BulleStyle(props), className, "up-bulle")}
+            style={{ backgroundImage: backgroundImage }}
+        >
+            <UpBox
+                flexDirection={"row"}
+                justifyContent={"flex-start"}
+                alignItems={"center"}
+            >
+                {icon &&
+                    <SvgIcon className={IconStyle} iconName={icon} color={'white'} />
+                }
+                {value &&
+                    <div className={ValueStyle}>{value}</div>
+                }
+                {message &&
+                    <div className={MessageStyle}>{message}</div>
+                }
+                {children &&
+                    <div className={ChildrenStyle}>{children}</div>
+                }
+            </UpBox>
+        </div>
+    )
+};
 
-    render() { 
-        return (
-            <div className={cn(BulleStyle(this.props), this.props.className, 'up-bulle')} style={{ backgroundImage: this.props.backgroundImage }}>
-                <UpBox flexDirection={'row'} justifyContent={'flex-start'} alignItems={'center'}>
-                    {this.props.icon &&
-                        <SvgIcon className={IconStyle} iconName={this.props.icon} color={"white"} />
-                    }
-                    {this.props.value &&
-                        <div className={ValueStyle}>{this.props.value}</div>
-                    }
-                    {this.props.message &&
-                        <div className={MessageStyle}>{this.props.message}</div>
-                    }
-                    {this.props.children &&
-                        <div className={ChildrenStyle}>{this.props.children}</div>
-                    }
-                </UpBox>
-            </div>
-        )
-    }
-}
+UpBulle.defaultProps = {
+    theme: defaultTheme
+};
 
-export default withTheme<UpBulleProps>(UpBulle)
+export { UpBulle };
+export default withTheme<UpBulleProps>(UpBulle);