+++ src/Components/Display/Bulle/types.ts
@@ -1,0 +1,9 @@
+import { IconName } from '../../../Common/theming/icons';
+
+export interface UpBulleProps {
+    backgroundImage: string;
+    message: string;
+    icon: IconName;
+    value: number;
+    className?: string;
+};