+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+#  ReactControls (version: 0.2.131)
+## Main Changes
+  - UpSvgIcon : Add warning icon
+
+#  ReactControls (version: 0.2.130)
+
+## Main Changes
+  - UpDataGrid : Change status of select all in case only one row can be selected
+
+
+#  ReactControls (version: 0.2.129)
+## Main Changes
+  - UpDataGrid : Change the behaviour for the mode only one row selectable
+  - Storybook : Upgrade to 6.1.17 
+#  ReactControls (version: 0.2.127)
+
+## Bug Fixes
+  - UpSelect : Fix No text message
+
+#  ReactControls (version: 0.2.126)
+
+## Main Changes
+  - UpText : Add max xharacters handling.
+
 #  ReactControls (version: 0.2.125)
 
 ## Bug Fixes